requirements.txt
Flask==3.0.3
gunicorn==21.2.0
google-generativeai==0.8.3
requests==2.32.3
